agents/creative_writer.py
```python
import os
from utils.openai_api import call_openai
from models.user_input import UserInputSchema
from models.project import ProjectSchema
from models.outline import OutlineSchema
from models.part import PartSchema
from models.step import StepSchema
from models.merge_verify import MergeVerifySchema

class CreativeWriter:
    def __init__(self):
        self.user_input = UserInputSchema()
        self.project = ProjectSchema()
        self.outline = OutlineSchema()
        self.part = PartSchema()
        self.step = StepSchema()
        self.merge_verify = MergeVerifySchema()

    def get_user_input(self):
        # Get user input for subject, keywords, target content type and word count
        self.user_input.get_input()

    def create_project(self):
        # Create a project plan based on user input
        self.project.create(self.user_input)

    def create_outline(self):
        # Create an outline based on the project plan
        self.outline.create(self.project)

    def define_part(self):
        # Define parts based on the outline
        self.part.define(self.outline)

    def generate_step(self):
        # Generate steps based on the defined parts
        self.step.generate(self.part)

    def merge_verify(self):
        # Merge and verify content against the user defined inputs
        self.merge_verify.merge(self.step)
        self.merge_verify.verify(self.user_input)

    def write_creative_content(self):
        # Get user input
        self.get_user_input()

        # Create project
        self.create_project()

        # Create outline
        self.create_outline()

        # Define parts
        self.define_part()

        # Generate steps
        self.generate_step()

        # Merge and verify
        self.merge_verify()

        # Return the final creative content
        return self.merge_verify.content

if __name__ == "__main__":
    writer = CreativeWriter()
    content = writer.write_creative_content()
    print(content)
```